Remove ResponsiveProp type and utils, optimize more arrays
